Outsource GPR code to resolution.py
